+++ test-tsd/tables.test-d.ts
@@ -78,7 +78,12 @@ const main = async () => {
     table.integer('num').references('id').withKeyName('non_for7').deferrable('deferred').inTable('non_exist').onDelete('CASCADE');
     table.integer('num').references('id').inTable('non_exist').onDelete('CASCADE').withKeyName('non_for6').deferrable('deferred');
     table.integer('num').references('id').withKeyName('non_for7').onDelete('CASCADE').deferrable('deferred').inTable('non_exist');
-    table.integer('num').references('id').withKeyName('non_for7').onDelete('CASCADE').deferrable('deferred').inTable('non_exist');
+
+    expectType<Knex.ReferencingColumnBuilder>(
+      table.integer('num').references('id').withKeyName('non_for7').onDelete('CASCADE')
+        .index('idx') // this shouldn't break type in chain
+        .deferrable('deferred').inTable('non_exist')
+    )
   })
 };