+++ test/integration/query/trigger-deletes.js
@@ -3,6 +3,15 @@
 const { expect } = require('chai');
 const { TEST_TIMESTAMP } = require('../../util/constants');
 const { isMssql } = require('../../util/db-helpers');
+const {
+  dropTables,
+  createAccounts,
+  createTestTableTwo,
+} = require('../../util/tableCreatorHelper');
+const {
+  insertAccounts,
+  insertTestTableTwoData,
+} = require('../../util/dataInsertHelper');
 
 module.exports = function (knex) {
   describe('Deletes with Triggers', function () {
@@ -13,6 +22,14 @@ module.exports = function (knex) {
       if (!isMssql(knex)) {
         this.skip('This test is MSSQL only');
       }
+    });
+
+    before(async () => {
+      await dropTables(knex);
+      await createAccounts(knex);
+      await createTestTableTwo(knex);
+      await insertAccounts(knex);
+      await insertTestTableTwoData(knex);
     });
 
     describe('Trigger Specific Tests', function () {
@@ -275,7 +292,7 @@ module.exports = function (knex) {
             'mssql',
             'select top(0) [t].* into #out from [test_table_two] as t left join [test_table_two] on 0=1;delete [test_table_two] output deleted.* into #out from [test_table_two] inner join [accounts] on [accounts].[id] = [test_table_two].[account_id] where [accounts].[email] = ?; select * from #out; drop table #out;',
             ['test3@example.com'],
-            [{ id: 3, account_id: 3, details: '', status: 1, json_data: null }]
+            [{ id: 3, account_id: 3, details: '', status: 1 }]
           );
         });
     });