+++ test/integration2/schema/foreign-keys.spec.js
@@ -1,16 +1,25 @@
-const { expect } = require('chai');
+const chai = require('chai');
+chai.use(require('chai-as-promised'));
+const expect = chai.expect;
 const sinon = require('sinon');
 const { getAllDbs, getKnexForDb } = require('../util/knex-instance-provider');
-const { isPostgreSQL, isSQLite } = require('../../util/db-helpers');
+const {
+  isPostgreSQL,
+  isSQLite,
+  isCockroachDB,
+} = require('../../util/db-helpers');
 
 describe('Schema', () => {
   describe('Foreign keys', () => {
     getAllDbs().forEach((db) => {
       describe(db, () => {
         let knex;
-        before(() => {
+        before(async () => {
           sinon.stub(Math, 'random').returns(0.1);
           knex = getKnexForDb(db);
+          await knex.schema.dropTableIfExists('foreign_keys_table_one');
+          await knex.schema.dropTableIfExists('foreign_keys_table_two');
+          await knex.schema.dropTableIfExists('foreign_keys_table_three');
         });
 
         after(() => {
@@ -34,10 +43,9 @@ describe('Schema', () => {
         });
 
         afterEach(async () => {
-          await knex.schema
-            .dropTable('foreign_keys_table_one')
-            .dropTable('foreign_keys_table_two')
-            .dropTable('foreign_keys_table_three');
+          await knex.schema.dropTable('foreign_keys_table_one');
+          await knex.schema.dropTable('foreign_keys_table_two');
+          await knex.schema.dropTable('foreign_keys_table_three');
         });
 
         describe('createForeignKey', () => {
@@ -138,14 +146,24 @@ describe('Schema', () => {
           it('creates new foreign key', async () => {
             await knex('foreign_keys_table_two').insert({});
             await knex('foreign_keys_table_three').insert({});
+
+            const rowsTwo = await knex('foreign_keys_table_two').select();
+            const rowsThree = await knex('foreign_keys_table_three').select();
+            const idTwo = rowsTwo[0].id;
+            const idThree = rowsThree[0].id;
+
             await knex('foreign_keys_table_one').insert({
-              fkey_two: 1,
-              fkey_three: 1,
+              fkey_two: idTwo,
+              fkey_three: idThree,
             });
             await knex('foreign_keys_table_one').insert({
-              fkey_two: 1,
-              fkey_three: 1,
+              fkey_two: idTwo,
+              fkey_three: idThree,
             });
+
+            const rowsOne = await knex('foreign_keys_table_one').select();
+            const idOne1 = rowsOne[0].id;
+            const idOne2 = rowsOne[1].id;
 
             await knex.schema.alterTable('foreign_keys_table_one', (table) => {
               table
@@ -157,22 +175,22 @@ describe('Schema', () => {
             const existingRows = await knex('foreign_keys_table_one').select();
             expect(existingRows).to.eql([
               {
-                fkey_three: 1,
-                fkey_two: 1,
-                id: 1,
+                fkey_three: idThree,
+                fkey_two: idTwo,
+                id: idOne1,
               },
               {
-                fkey_three: 1,
-                fkey_two: 1,
-                id: 2,
+                fkey_three: idThree,
+                fkey_two: idTwo,
+                id: idOne2,
               },
             ]);
 
             await knex('foreign_keys_table_two').insert({});
             await knex('foreign_keys_table_three').insert({});
             await knex('foreign_keys_table_one').insert({
-              fkey_two: 1,
-              fkey_three: 1,
+              fkey_two: idTwo,
+              fkey_three: idThree,
             });
             try {
               await knex('foreign_keys_table_one').insert({
@@ -189,6 +207,11 @@ describe('Schema', () => {
               if (isPostgreSQL(knex)) {
                 expect(err.message).to.equal(
                   `insert into "foreign_keys_table_one" ("fkey_three", "fkey_two") values ($1, $2) - insert or update on table "foreign_keys_table_one" violates foreign key constraint "fk_fkey_threeee"`
+                );
+              }
+              if (isCockroachDB(knex)) {
+                expect(err.message).to.equal(
+                  `insert into "foreign_keys_table_one" ("fkey_three", "fkey_two") values ($1, $2) - insert on table "foreign_keys_table_one" violates foreign key constraint "fk_fkey_threeee"`
                 );
               }
               expect(err.message).to.include('constraint');
@@ -240,6 +263,107 @@ describe('Schema', () => {
               `PRAGMA foreign_key_list('foreign_keys_table_one');`
             );
             expect(fks.length).to.equal(1);
+          });
+        });
+
+        describe('Schema (Foreign keys)', () => {
+          beforeEach(async () => {
+            await knex.schema
+              .dropTableIfExists('foreign_keys_table_one')
+              .dropTableIfExists('foreign_keys_table_four');
+
+            await knex.schema
+              .createTable('foreign_keys_table_four', (table) => {
+                table.string('col1').notNull();
+                table.string('col2').notNull();
+                table.primary(['col1', 'col2']);
+              })
+              .createTable('foreign_keys_table_one', (table) => {
+                table.increments();
+                table.integer('fkey_two').unsigned().notNull();
+                table
+                  .foreign('fkey_two')
+                  .references('foreign_keys_table_two.id');
+                table.string('fkey_four_part1');
+                table.string('fkey_four_part2');
+                table
+                  .foreign(['fkey_four_part1', 'fkey_four_part2'])
+                  .references(['col1', 'col2'])
+                  .inTable('foreign_keys_table_four');
+                table.integer('fkey_three').unsigned().notNull();
+                table
+                  .foreign('fkey_three')
+                  .references('foreign_keys_table_three.id')
+                  .withKeyName('fk_fkey_threeee');
+              });
+          });
+          afterEach(async () => {
+            await knex.schema.dropTableIfExists('foreign_keys_table_four');
+          });
+
+          describe('drop foreign key', () => {
+            it('correctly drops foreign key', async () => {
+              await knex('foreign_keys_table_two').insert({});
+              await knex('foreign_keys_table_three').insert({});
+              await knex('foreign_keys_table_four').insert({
+                col1: 'a',
+                col2: 'b',
+              });
+              const tableTwoEntry = await knex(
+                'foreign_keys_table_two'
+              ).select();
+              const tableThreeEntry = await knex(
+                'foreign_keys_table_three'
+              ).select();
+              await knex('foreign_keys_table_one').insert({
+                fkey_two: tableTwoEntry[0].id,
+                fkey_three: tableThreeEntry[0].id,
+                fkey_four_part1: 'a',
+                fkey_four_part2: 'b',
+              });
+              try {
+                await knex('foreign_keys_table_one').insert({
+                  fkey_two: 9999,
+                  fkey_three: 99,
+                  fkey_four_part1: 'a',
+                  fkey_four_part2: 'b',
+                });
+                throw new Error("Shouldn't reach this");
+              } catch (err) {
+                if (isSQLite(knex)) {
+                  expect(err.message).to.equal(
+                    `insert into \`foreign_keys_table_one\` (\`fkey_four_part1\`, \`fkey_four_part2\`, \`fkey_three\`, \`fkey_two\`) values ('a', 'b', 99, 9999) - SQLITE_CONSTRAINT: FOREIGN KEY constraint failed`
+                  );
+                }
+                if (isPostgreSQL(knex)) {
+                  expect(err.message).to.equal(
+                    `insert into "foreign_keys_table_one" ("fkey_four_part1", "fkey_four_part2", "fkey_three", "fkey_two") values ($1, $2, $3, $4) - insert or update on table "foreign_keys_table_one" violates foreign key constraint "foreign_keys_table_one_fkey_two_foreign"`
+                  );
+                }
+                if (isCockroachDB(knex)) {
+                  expect(err.message).to.equal(
+                    `insert into "foreign_keys_table_one" ("fkey_four_part1", "fkey_four_part2", "fkey_three", "fkey_two") values ($1, $2, $3, $4) - insert on table "foreign_keys_table_one" violates foreign key constraint "foreign_keys_table_one_fkey_two_foreign"`
+                  );
+                }
+                expect(err.message).to.include('constraint');
+              }
+
+              await knex.schema.alterTable(
+                'foreign_keys_table_one',
+                (table) => {
+                  table.dropForeign(['fkey_two']);
+                  table.dropForeign([], 'fk_fkey_threeee');
+                  table.dropForeign(['fkey_four_part1', 'fkey_four_part2']);
+                }
+              );
+
+              await knex('foreign_keys_table_one').insert({
+                fkey_two: 999,
+                fkey_three: 999,
+                fkey_four_part1: 'e',
+                fkey_four_part2: 'f',
+              });
+            });
           });
         });
       });