+++ test/integration2/schema/primary-keys.spec.js
@@ -1,5 +1,10 @@
 const { expect } = require('chai');
-const { isMssql, isSQLite, isPostgreSQL } = require('../../util/db-helpers');
+const {
+  isMssql,
+  isSQLite,
+  isPgBased,
+  isCockroachDB,
+} = require('../../util/db-helpers');
 const { getAllDbs, getKnexForDb } = require('../util/knex-instance-provider');
 
 describe('Schema', () => {
@@ -10,9 +15,6 @@ describe('Schema', () => {
 
         before(function () {
           knex = getKnexForDb(db);
-          if (isMssql(knex)) {
-            return this.skip();
-          }
         });
 
         after(() => {
@@ -34,7 +36,7 @@ describe('Schema', () => {
         describe('createPrimaryKey', () => {
           it('creates a new primary key', async () => {
             await knex.schema.alterTable('primary_table', (table) => {
-              table.integer('id_four').primary();
+              table.integer('id_four').notNull().primary();
             });
 
             await knex('primary_table').insert({ id_four: 1 });
@@ -48,7 +50,7 @@ describe('Schema', () => {
                   'insert into `primary_table` (`id_four`) values (1) - SQLITE_CONSTRAINT: UNIQUE constraint failed: primary_table.id_four'
                 );
               }
-              if (isPostgreSQL(knex)) {
+              if (isPgBased(knex)) {
                 expect(err.message).to.equal(
                   'insert into "primary_table" ("id_four") values ($1) - duplicate key value violates unique constraint "primary_table_pkey"'
                 );
@@ -56,9 +58,17 @@ describe('Schema', () => {
             }
           });
 
-          it('creates a primary key with a custom constraint name', async () => {
+          it('creates a primary key with a custom constraint name', async function () {
+            // CockroachDB 21.1 throws "(72): unimplemented: primary key dropped without subsequent addition of new primary key in same transaction"
+            if (isCockroachDB(knex)) {
+              return this.skip();
+            }
+
             await knex.schema.alterTable('primary_table', (table) => {
-              table.integer('id_four').primary('my_custom_constraint_name');
+              table
+                .integer('id_four')
+                .notNull()
+                .primary('my_custom_constraint_name');
             });
 
             await knex('primary_table').insert({ id_four: 1 });
@@ -71,7 +81,7 @@ describe('Schema', () => {
                   'insert into `primary_table` (`id_four`) values (1) - SQLITE_CONSTRAINT: UNIQUE constraint failed: primary_table.id_four'
                 );
               }
-              if (isPostgreSQL(knex)) {
+              if (isPgBased(knex)) {
                 expect(err.message).to.equal(
                   'insert into "primary_table" ("id_four") values ($1) - duplicate key value violates unique constraint "my_custom_constraint_name"'
                 );
@@ -87,6 +97,11 @@ describe('Schema', () => {
 
           it('creates a compound primary key', async () => {
             await knex.schema.alterTable('primary_table', (table) => {
+              // CockroachDB and mssql do not support nullable primary keys
+              if (isCockroachDB(knex) || isMssql(knex)) {
+                table.dropNullable('id_two');
+                table.dropNullable('id_three');
+              }
               table.primary(['id_two', 'id_three']);
             });
 
@@ -102,7 +117,7 @@ describe('Schema', () => {
                   'insert into `primary_table` (`id_three`, `id_two`) values (1, 1) - SQLITE_CONSTRAINT: UNIQUE constraint failed: primary_table.id_two, primary_table.id_three'
                 );
               }
-              if (isPostgreSQL(knex)) {
+              if (isPgBased(knex)) {
                 expect(err.message).to.equal(
                   'insert into "primary_table" ("id_three", "id_two") values ($1, $2) - duplicate key value violates unique constraint "primary_table_pkey"'
                 );
@@ -110,39 +125,54 @@ describe('Schema', () => {
             }
           });
 
-          it('creates a compound primary key with a custom constraint name', async () => {
-            await knex.schema.alterTable('primary_table', (table) => {
-              table.primary(
-                ['id_two', 'id_three'],
-                'my_custom_constraint_name'
-              );
-            });
+          for (const [flavor, customConstraintName] of [
+            ['provided directly as a string', 'my_custom_constraint_name'],
+            [
+              'provided in the options object',
+              { constraintName: 'my_custom_constraint_name' },
+            ],
+          ]) {
+            it(`creates a compound primary key with a custom constraint name ${flavor}`, async function () {
+              // As of 2021-10-02, CockroachDB does not support dropping a primary key without creating a new one in the same transaction.
+              if (isCockroachDB(knex)) {
+                return this.skip();
+              }
 
-            await knex('primary_table').insert({ id_two: 1, id_three: 1 });
-            await knex('primary_table').insert({ id_two: 2, id_three: 1 });
-            await knex('primary_table').insert({ id_two: 1, id_three: 2 });
+              await knex.schema.alterTable('primary_table', (table) => {
+                // CockroachDB and mssql do not support nullable primary keys
+                if (isCockroachDB(knex) || isMssql(knex)) {
+                  table.dropNullable('id_two');
+                  table.dropNullable('id_three');
+                }
+                table.primary(['id_two', 'id_three'], customConstraintName);
+              });
 
-            try {
               await knex('primary_table').insert({ id_two: 1, id_three: 1 });
-            } catch (err) {
-              if (isSQLite(knex)) {
-                expect(err.message).to.equal(
-                  'insert into `primary_table` (`id_three`, `id_two`) values (1, 1) - SQLITE_CONSTRAINT: UNIQUE constraint failed: primary_table.id_two, primary_table.id_three'
-                );
-              }
-              if (isPostgreSQL(knex)) {
-                expect(err.message).to.equal(
-                  'insert into "primary_table" ("id_three", "id_two") values ($1, $2) - duplicate key value violates unique constraint "my_custom_constraint_name"'
-                );
-              }
-            }
+              await knex('primary_table').insert({ id_two: 2, id_three: 1 });
+              await knex('primary_table').insert({ id_two: 1, id_three: 2 });
 
-            await knex.schema.alterTable('primary_table', (table) => {
-              table.dropPrimary('my_custom_constraint_name');
+              try {
+                await knex('primary_table').insert({ id_two: 1, id_three: 1 });
+              } catch (err) {
+                if (isSQLite(knex)) {
+                  expect(err.message).to.equal(
+                    'insert into `primary_table` (`id_three`, `id_two`) values (1, 1) - SQLITE_CONSTRAINT: UNIQUE constraint failed: primary_table.id_two, primary_table.id_three'
+                  );
+                }
+                if (isPgBased(knex)) {
+                  expect(err.message).to.equal(
+                    'insert into "primary_table" ("id_three", "id_two") values ($1, $2) - duplicate key value violates unique constraint "my_custom_constraint_name"'
+                  );
+                }
+              }
+
+              await knex.schema.alterTable('primary_table', (table) => {
+                table.dropPrimary('my_custom_constraint_name');
+              });
+
+              await knex('primary_table').insert({ id_two: 1, id_three: 1 });
             });
-
-            await knex('primary_table').insert({ id_two: 1, id_three: 1 });
-          });
+          }
         });
       });
     });