+++ test/util/db-helpers.js
@@ -18,7 +18,21 @@ function isPgBased(knex) {
     drivers.PostgreSQL,
     drivers.PgNative,
     drivers.Redshift,
+    drivers.CockroachDB,
   ]);
+}
+
+function isPgBasedDriverName(driverName) {
+  return [
+    drivers.PostgreSQL,
+    drivers.PgNative,
+    drivers.Redshift,
+    drivers.CockroachDB,
+  ].includes(driverName);
+}
+
+function isPgDriverName(driverName) {
+  return [drivers.PostgreSQL, drivers.PgNative].includes(driverName);
 }
 
 function isMssql(knex) {
@@ -41,10 +55,14 @@ function isSQLite(knex) {
   return getDriverName(knex) === drivers.SQLite;
 }
 
+function isCockroachDB(knex) {
+  return getDriverName(knex) === drivers.CockroachDB;
+}
+
 /**
  *
  * @param knex
- * @param {('pg'|'pgnative'|'pg-redshift'|'oracledb'|'mysql'|'mysql2'|'mssql'|'sqlite3')[]} supportedDbs - supported DB values in DRIVER_NAMES from lib/constants.
+ * @param {('pg'|'pgnative'|'pg-redshift'|'oracledb'|'mysql'|'mysql2'|'mssql'|'sqlite3'|'cockroachdb')[]} supportedDbs - supported DB values in DRIVER_NAMES from lib/constants.
  * @returns {*}
  */
 function isOneOfDbs(knex, supportedDbs) {
@@ -53,12 +71,15 @@ function isOneOfDbs(knex, supportedDbs) {
 
 module.exports = {
   isOneOfDbs,
+  isCockroachDB,
   isMysql,
   isMssql,
   isOracle,
   isPostgreSQL,
   isPgNative,
+  isPgDriverName,
   isPgBased,
+  isPgBasedDriverName,
   isRedshift,
   isSQLite,
 };